+++ spec/spec_helper.rb
@@ -1,6 +1,4 @@
-require 'rubygems'
 require 'ruby_speech'
-require 'mocha'
 
 include RubySpeech::GRXML
 include RubySpeech::XML::Language
@@ -16,7 +14,6 @@ schema_file_path = File.expand_path File.join(__FILE__, '../../assets/grammar.xs
 puts "Loading the GRXML Schema from #{schema_file_path}..."
 GRXML_SCHEMA = Nokogiri::XML::Schema File.open(schema_file_path)
 puts "Finished loading schema."
-
 
 RSpec.configure do |config|
   config.mock_with :mocha